+++ pyogrio/geopandas.py
@@ -124,7 +124,7 @@ def read_dataframe(
         installed). When enabled, this provides a further speed-up.
     **kwargs
         Additional driver-specific dataset open options passed to OGR.  Invalid
-        options are logged by OGR to stderr and are not captured.
+        options will trigger a warning.
 
     Returns
     -------